Add files via upload
I was unable to figure out how to import the shapes library in order to properly output a rectangle to manipulate. I looked up various sets of codes, including trying the code that was sent via email, but it would not run in my sublime. I am confused about how to call in the shapes library and the order of the code, as I keep getting red flags stating that there is a syntax error.