Add files via upload
This is the final product of my final project code. 
For my project I gathered COVID-19 information for five counties in my area and compared the population of each county to the percentage of those who were tested for COVID-19, positive for COVID-19, and the those who died due to COVID-19.  I displayed a multi-colored bar graph that compares the percentages of those tested, positive for infection, and deaths in each county. These numbers were taken from the Johns Hopkins COVID-19 map/website, and the numbers were from statistics from about two weeks or so ago.